Upgrade to release 20120329.

+++ trunk/discuzx2.0ML/upload/source/include/post/post_editpost.php
@@ -4,7 +4,7 @@
  *      [Discuz!] (C)2001-2099 Comsenz Inc.
  *      This is NOT a freeware, use is subject to license terms
  *
- *      $Id: post_editpost.php 25514 2011-11-14 02:37:37Z monkey $
+ *      $Id: post_editpost.php 29004 2012-03-22 04:04:58Z chenmengshu $
  *	Modified by Valery Votintsev, codersclub.org
  */
 
@@ -616,16 +616,19 @@ if(!submitcheck('editsubmit')) {
 				DB::query("UPDATE ".DB::table('forum_threadrush')." SET stopfloor='$_G[gp_stopfloor]', starttimefrom='$_G[gp_rushreplyfrom]', starttimeto='$_G[gp_rushreplyto]', rewardfloor='$_G[gp_rewardfloor]' WHERE tid='$_G[tid]'", 'UNBUFFERED');
 			}
 //vot escape_str added
-			DB::query("UPDATE ".DB::table('forum_thread')."
-                                            SET typeid='$typeid', sortid='$sortid',
-                                            subject='".DB::escape_str($subject)."',
-                                            readperm='$readperm', price='$price' $closedadd $authoradd $polladd $replycreditadd".($_G['forum_auditstatuson'] && $audit == 1 ? ",displayorder='0', moderated='1'" : ",displayorder='$displayorder'").", status='$thread[status]' WHERE tid='$_G[tid]'", 'UNBUFFERED');
-//vot escape_str added
-			$_G['tid'] > 1 && DB::query("UPDATE ".DB::table('forum_thread')."
-                                           SET subject='".DB::escape_str($subject)."'
-                                           WHERE closed='$_G[tid]'", 'UNBUFFERED');
 
-			$tagstr = modthreadtag($_G['gp_tags'], $_G[tid]);
+			DB::query("UPDATE ".DB::table('forum_thread')."
+                                   SET typeid='$typeid',
+					sortid='$sortid',
+                                        subject='".DB::escape_str($subject)."',
+                                        readperm='$readperm', price='$price' $closedadd $authoradd $polladd $replycreditadd".($_G['forum_auditstatuson'] && $audit == 1 ? ",displayorder='0', moderated='1'" : ",displayorder='$displayorder'").", status='$thread[status]'
+				   WHERE tid='$_G[tid]'", 'UNBUFFERED');
+//vot escape_str added
+			$thread['closed'] > 1 && DB::query("UPDATE ".DB::table('forum_thread')."
+                                           SET subject='".DB::escape_str($subject)."'
+                                           WHERE tid='$thread[closed]'", 'UNBUFFERED');
+
+			$tagstr = modthreadtag($_G['gp_tags'], $_G['tid']);
 
 		} else {
 
@@ -723,11 +726,13 @@ if(!submitcheck('editsubmit')) {
 
 		if($special == 4 && $isfirstpost && $_G['group']['allowpostactivity']) {
 			$activityaid = DB::result_first("SELECT aid FROM ".DB::table('forum_activity')." WHERE tid='$_G[tid]'");
-			if($activityaid != $_G['gp_activityaid']) {
+			if($activityaid && $activityaid != $_G['gp_activityaid']) {
 				$attach = DB::fetch_first("SELECT attachment, thumb, remote, aid FROM ".DB::table(getattachtablebytid($_G['tid']))." WHERE aid='$activityaid'");
 				DB::query("DELETE FROM ".DB::table('forum_attachment')." WHERE aid='$activityaid'");
 				DB::query("DELETE FROM ".DB::table(getattachtablebytid($_G['tid']))." WHERE aid='$activityaid'");
 				dunlink($attach);
+			}
+			if($_G['gp_activityaid']) {
 				$threadimageaid = $_G['gp_activityaid'];
 				convertunusedattach($_G['gp_activityaid'], $_G['tid'], $pid);
 				DB::query("UPDATE ".DB::table('forum_activity')." SET aid='$_G[gp_activityaid]' WHERE tid='$_G[tid]'");
@@ -991,4 +996,3 @@ if(!submitcheck('editsubmit')) {
 
 }
 
-?>
